+++ examples/simple-chart/index.js
@@ -1,3 +1,5 @@
+/// <reference path="../../packages/d3fc/index.d.ts" />
+
 const data = fc.randomFinancial()(50);
 
 const yExtent = fc.extentLinear().accessors([d => d.high, d => d.low]);
@@ -10,9 +12,10 @@ const multi = fc.seriesSvgMulti().series([gridlines, candlestick]);
 
 const chart = fc
     .chartCartesian(d3.scaleTime(), d3.scaleLinear())
-    .yDomain(yExtent(data))
-    .xDomain(xExtent(data))
     .svgPlotArea(multi);
+
+chart.xDomain(xExtent(data));
+chart.yDomain(yExtent(data));
 
 d3.select('#chart')
     .datum(data)