+++ packages/d3fc/index.d.ts
@@ -1,0 +1,170 @@
+export as namespace fc;
+
+export * from '@d3fc/d3fc-chart';
+
+// ---------------------------------------------------------------------------------
+// The following types are placeholders until the package has real definitions added
+// ---------------------------------------------------------------------------------
+
+// @d3fc/d3fc-annotation
+export const annotationCanvasBand: any;
+export const annotationCanvasCrosshair: any;
+export const annotationCanvasGridline: any;
+export const annotationCanvasLine: any;
+export const annotationSvgBand: any;
+export const annotationSvgCrosshair: any;
+export const annotationSvgGridline: any;
+export const annotationSvgLine: any;
+
+// @d3fc/d3fc-axis
+export const axisBottom: any;
+export const axisLabelOffset: any;
+export const axisLabelRotate: any;
+export const axisLeft: any;
+export const axisOrdinalBottom: any;
+export const axisOrdinalLeft: any;
+export const axisOrdinalRight: any;
+export const axisOrdinalTop: any;
+export const axisRight: any;
+export const axisTop: any;
+
+// @d3fc/d3fc-brush
+export const brush: any;
+export const brushX: any;
+export const brushY: any;
+
+// @d3fc/d3fc-data-join
+export const dataJoin: any;
+export const effectivelyZero: any;
+export const isTransition: any;
+
+// @d3fc/d3fc-discontinuous-scale
+export const discontinuityIdentity: any;
+export const discontinuityRange: any;
+export const discontinuitySkipUtcWeekends: any;
+export const discontinuitySkipWeekends: any;
+export const scaleDiscontinuous: any;
+
+// @d3fc/d3fc-extent
+export const extentLinear: any;
+export const extentTime: any;
+
+// @d3fc/d3fc-financial-feed
+export const feedGdax: any;
+
+// @d3fc/d3fc-group
+export const group: any;
+
+// @d3fc/d3fc-technical-indicator
+export const indicatorBollingerBands: any;
+export const indicatorElderRay: any;
+export const indicatorEnvelope: any;
+export const indicatorExponentialMovingAverage: any;
+export const indicatorForceIndex: any;
+export const indicatorMacd: any;
+export const indicatorMovingAverage: any;
+export const indicatorRelativeStrengthIndex: any;
+export const indicatorStochasticOscillator: any;
+
+// @d3fc/d3fc-label-layout
+export const layoutAnnealing: any;
+export const layoutBoundingBox: any;
+export const layoutGreedy: any;
+export const layoutLabel: any;
+export const layoutRemoveOverlaps: any;
+export const layoutTextLabel: any;
+
+// @d3fc/d3fc-pointer
+export const pointer: any;
+
+// @d3fc/d3fc-random
+export const randomFinancial: any;
+export const randomGeometricBrownianMotion: any;
+export const randomSkipWeekends: any;
+
+// @d3fc/d3fc-rebind
+export const exclude: any;
+export const include: any;
+export const includeMap: any;
+export const prefix: any;
+export const rebind: any;
+export const rebindAll: any;
+
+// @d3fc/d3fc-sample
+export const bucket: any;
+export const modeMedian: any;
+export const largestTriangleOneBucket: any;
+export const largestTriangleThreeBucket: any;
+
+// @d3fc/d3fc-series
+export const autoBandwidth: any;
+export const seriesCanvasArea: any;
+export const seriesCanvasBar: any;
+export const seriesCanvasBoxPlot: any;
+export const seriesCanvasCandlestick: any;
+export const seriesCanvasErrorBar: any;
+export const seriesCanvasGrouped: any;
+export const seriesCanvasHeatmap: any;
+export const seriesCanvasLine: any;
+export const seriesCanvasMulti: any;
+export const seriesCanvasOhlc: any;
+export const seriesCanvasPoint: any;
+export const seriesCanvasRepeat: any;
+export const seriesSvgArea: any;
+export const seriesSvgBar: any;
+export const seriesSvgBoxPlot: any;
+export const seriesSvgCandlestick: any;
+export const seriesSvgErrorBar: any;
+export const seriesSvgGrouped: any;
+export const seriesSvgHeatmap: any;
+export const seriesSvgLine: any;
+export const seriesSvgMulti: any;
+export const seriesSvgOhlc: any;
+export const seriesSvgPoint: any;
+export const seriesSvgRepeat: any;
+export const seriesWebglArea: any;
+export const seriesWebglBar: any;
+export const seriesWebglBoxPlot: any;
+export const seriesWebglCandlestick: any;
+export const seriesWebglErrorBar: any;
+export const seriesWebglLine: any;
+export const seriesWebglMulti: any;
+export const seriesWebglOhlc: any;
+export const seriesWebglPoint: any;
+export const seriesWebglRepeat: any;
+
+// @d3fc/d3fc-shape
+export const shapeBar: any;
+export const shapeBoxPlot: any;
+export const shapeCandlestick: any;
+export const shapeErrorBar: any;
+export const shapeOhlc: any;
+
+// @d3fc/d3fc-webgl
+export const webglAdjacentAttribute: any;
+export const webglAttribute: any;
+export const webglBaseAttribute: any;
+export const webglBufferBuilder: any;
+export const webglElementIndices: any;
+export const webglFillColor: any;
+export const webglProgramBuilder: any;
+export const webglScaleLinear: any;
+export const webglScaleLog: any;
+export const webglScaleMapper: any;
+export const webglScalePow: any;
+export const webglSeriesArea: any;
+export const webglSeriesBar: any;
+export const webglSeriesBoxPlot: any;
+export const webglSeriesCandlestick: any;
+export const webglSeriesErrorBar: any;
+export const webglSeriesLine: any;
+export const webglSeriesOhlc: any;
+export const webglSeriesPoint: any;
+export const webglShaderBuilder: any;
+export const webglStrokeColor: any;
+export const webglSymbolMapper: any;
+export const webglTypes: any;
+export const webglUniform: any;
+
+// @d3fc/d3fc-zoom
+export const zoom: any;